requirements.txt
streamlit==1.37.1
streamlit
torch
torchvision
pillow
numpy